+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog file for dinero-rs project, a command line application for managing fi
 ## [0.32.0] - xxx
 ### Added
 - Implemented ```date-format```
+- Added ```--calendar``` to the ```roi``` command, showing a [calendar view of TWR](https://github.com/frosklis/dinero-rs/issues/115).
+- Added ```--no-summary``` flag to the ```roi``` command, to suppress the summary after the table
 ### Fixed
 - [```args-only```](https://github.com/frosklis/dinero-rs/issues/120)