+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
-## [0.32.0] - xxx
+## [0.32.0] - 2021-08-24
 ### Added
 - Implemented ```date-format```
 - Added ```--calendar``` to the ```roi``` command, showing a [calendar view of TWR](https://github.com/frosklis/dinero-rs/issues/115).